Updated Create_certificate to key of length 2048
deb package creates a key of length 1024 by default and and since the release of openssl 1.1.1 the default openssl.conf file at /etc/ssl/openssl.conf  has CipherString = DEFAULT@SECLEVEL=2 , resulting in the key length to be small.  In order to adhere to SECLEVEL=2 the key length was increased from 1024 to 2048.