Use addopts instead of testpaths to specify path in pytest config
The pytest documentation states that the testpaths configuration option
"Sets list of directories that should be searched ... when executing
pytest from the rootdir directory." Since pytest is not executed from
the rootdir directory, testpaths has no effect. Appending the "tests/"
directory to the end of addopts reduces the time required to run the
test suite by approximately 6 seconds.