Remove "__author__" lines from the source code
The "__author__" lines in numerous Python files serve little purpose.
At best, they are redundant. At worst, they are incorrect and
misleading.

The VCS (i.e. git) tracks all changes to files. Storing the author in
the files is redundant violates the DRY principle.

If the file changes, whose name belongs in the __author__ field? What if
it changes so drastically that it's almost unrecognizable? If there's
any circumstance where the __author__ name should change, what enforces
that this change takes place? Given the ambiguity of these scenarios,
the __author__ field could be misleading or incorrect.

Resolves #1345