Tests: Skip batching telemetry tests on Windows
The batching telemetry tests do not function properly on Windows. They
sometimes pass and sometimes fail due to a race condition. I've been
unable to get this race condition to manifest on Linux and I believe
this is likely due to differences in how threads and the GIL are handled
between the two operating systems.