cc: Add sleep function & call it when setting the scenario
This is done to give a small time gap between sending the selected mode
to the backend in `LandingPage.tsx`'s `setScenario`, and fetching and
setting the mode in the frontend in `Main.tsx`'s `setMode()`.
Without the small time gap, the backend is not able to register
the selected mode in time for the frontend to fetch the selected mode.
This causes a big gap (at most 10 seconds) when nothing
happens on the frontend until the periodic `updateStatus()` is called
(which calls `setMode()`).