UI: Fix missing default brute-forcer options in config schema
If we don't provide default options in config schema AdvancedMultiSelect doesn't know what to reset it back to. Essentially we need two states, one for the actual value of configuration and one for the default. When we disable SMB, the SMB value with all it's options are deleted. When we enable it, we need to pull its options from somewhere.