Agent: Supress AssertionError when unlocking the singleton
When unlock() is called, it asserts that the singleton is locked. If the
singleton could not be acquired, then the call to unlock() will raise an
exception. In our finally clause, we don't care whether or not we're
unlocking a locked singleton or if unlock() fails because the singleton
was not acquired, so we can suppress AssertionError.