Agent: Define SOCKET_TIMEOUT in terms of HEARTBEAT_INTERVAL
The previous SOCKET_TIMEOUT was chosen because it "seemed reasonable".
Since there was no criteria by which any component could determine
whether or not an agent was still alive, 10 seconds was as good a guess
as any.

Since the introduction of the Agent's heartbeat (Issue #2518),
components can make reasonable assumptions about the status of an agent
based on whether or not they've received a heartbeat.

If the socket hasn't received any data from the agent within 2 heartbeat
intervals, it's probably safe to assume the agent is dead and close the
socket.

Fixes #2635
PR #2652