Agent: Remove Thread inheritance from PluginEventForwarder
Per the Python documentation, "only override the __init__() and run()
methods of this class [Thread]."

In order to start and stop PluginEventForwarder, start() must clear()
the internal self._stop() Event. Since we can't override start() in a
Thread child class, `PluginEventForwarder` must not inherit from Thread.
Additionally, this makes it easier to change concurrency mechanisms in
the future.