Agent: Set reset_modules_cache=False when loading MultiprocessingPlugin
Pydantic is an extension module and is loaded into `sys.modules` when
the child process starts. Setting `reset_modules_cache=False` avoids the
issues associated with attempting to unload extension modules.

It's not clear why pydantic is in sys.modules when the child process
starts, but this change will keep development moving until we can
investigate more thoroughly.