Agent: Merge "brute_force" and "vulnerability" into "exploiters"
The separation between brute force and vulnerability exploiters was
based on the use case of "vulnerability scanning", which is no longer
Infection Monkey's main goal. Ultimately, users should be able to
configure the order that exploiters run in. While the UI doesn't allow
this at the moment, the configuration can be manipulated manually to
achieve this. However, vulnerability exploiters always run before brute
force exploiters. In other words, this separation reduces the
configurability of the system.

By merging these into a single list of exploiters, we allow the user to
have total control over the order in which exploiters are run.

Issue #2787
PR #2789