Agent: Spawn the manager process before acquiring the SystemSingleton
The SystemSingleton uses an abstract unix socket as a "lock" to ensure
only one agent at a time runs on a given machine. It seems that if a
manager process is spawned after this unix socket is created, the
manager process inherits this file handle, which leads to the socket
never being properly closed.

Spawning the manager before the socket is opened is a quick solution to
this problem. A better solution (see
https://github.com/guardicore/monkey/issues/2817) is to use a different
method than a unix socket to achieve this goal, but, baby steps for now.