Agent: Forcefully kill the LDAP server process if necessary
Using Twisted for the Log4Shell exploiter has been nothing but trouble
since the beginning. When we refactor this exploiter we should use
another solution. In the meanwhile, we must be doing something wrong WRT
stopping Twisted. The heavy-handed approach is to SIGKILL the process.
This isn't ideal, but will be changed when we refactor this component.

Issue #2820