Hadoop: Remove errant setting of ExploiterResultData.os
The agent scans network nodes and makes assumptions about what OS is
running on them. The exploiter, on the other hand, may be able to have
much more accurate knowledge since it can run code on the machine. The
purpose of ExploiterResultData.os is to override the scanner's
assumptions, not to simply contain the OS from the `TargetHost` passed
to the exploiter.