Agent: Handle scan results
The Propagator does not expect previous TCP scan results, so it will
"merge" the scan data by overwriting old scan data, in cases where there
might be conflicting scan data for a given port.

The fingerprint data is merged in a similar manner, but the service is
only overwritten if the new service information is more descriptive than
the existing info. Currently this means that the service is only updated
if the new service is not UNKNOWN.