Hadoop: Remove "export" from Linux command
Using `export` in agent run commands is unnecessary and undesirable. It
will add the environment variable to the parent process's environment,
not just the environment of the agent process. This will expose the OTP
more than necessary, violating the principle of least privilege. It will
also make it difficult (impossible?) for the agent to properly clean up
all traces of the OTP.