Common: Make OTP TypeAlias a SecretStr
The `ISecretVariable` was added to attempt to isolate other components
from pydantic. While a noble attempt, the interface is too loose to be
truly valuable (get_secret_value() returns Any). The OTP type alias
itself serves the purpose of isolating the code from pydantic. The only
pydantic "leakage" would be the name of the method,
`get_secret_value()`. However, if we removed pydantic, as long as
whatever replaced `pydantic.SecretStr` still provides
`get_secret_value() -> str`, no other code needs to change.

We could provide `ISecretStr`, but if its interface is identical to
`SecretStr`, I don't see what value that would offer us.

An additional note: We'd prefer OTP to be a concrete type so that
strings can be converted to OTPs easily with `OTP(my_otp_str)`.