Agent: Add `asyncore` as a hidden import in monkey.spec
The SNMP plugin depends on the `asyncore` package, which is part of the
standard library. Our plugins assume that standard library modules are
available. These modules are made available to the plugins in the agent
executable. However, pyinstaller does not include modules that are not
explicitly imported, even if they're part of the standard library. Since
`asyncore` is not used in the agent, pyinstaller is not pulling in that
module. Therefore, we need to tell pyinstaller to include it, which we
do by adding it to the hidden imports in the spec file.