SNMP: Change how exploiter short-circuits brute-force
Use a simple command to check if SNMP is running prior to attempting to
brute-force with community strings. The previous method of
short-circuiting on timeout was insufficient because an SNMP request
with an invalid community string results in a timeout, which would
short-circuit brute-forcing and prevent exploitation if an invalid
community string was attepted before a valid community string.