Island: Remove caching from MasqueradeAgentBinaryRepositoryDecorator
MasqueradeAgentBinaryRepositoryDecorator relies on the
IMasqueradeRepository and has no knowledge of whether or not the
contents of that repository have changed. As such, it cannot cache
results, as it does not know whether or not it should retreive cached
results or regenerate them.

There is, conceivably, a different caching strategy that could work for
this component. Until this component's performance becomes an issue,
we'll just skip the caching for now.