Agent: Finish cleanup if Island is unreachable
It's possible that the agent is shutting down because the Island has
become unreachable. If this is the case, calls to `should_agent_stop()`
will raise an exception, causing a number of cleanup operations in
`cleanup()` to be skipped.

This commit wraps these calls in a try/except to prevent
short-circuiting the cleanup process.