Cryptojacker: Mark cryptojacker as unsafe for production
Since this payload consumes resources, it's possible that this could
cause servers or the processes running on them to crash, become
unresponsive, or respond unacceptably slow. This depends on a variety of
factors which the plugin itself cannot control. Since the safety of this
plugin depends upon the environment and how the user configures it, I'm
marking it unsafe.