Common: Implement monkeytypes facade
A number of types and models have been broken out of common/ and put in
a new, standalone library named monkeytypes. This commit modifies common
to behave as a facade for monkeytypes, allowing the Island, Agent, and
plugins to be unchanged by this transition. Each of those components
will be individually migrated to monkeytypes in future commits. This
facade will then be removed from common once there are no more
references to it.