BB: Use a keep_tunnel_open_time > 0 for depth_2 test
It appears that will a value of 0, an agent providing a relay could
sometimes close before a propagated agent was able to fully establish
communication with the Island. In those instances, the propagated agent
could successfully query the Island status, but after the 5s timeout
lapsed testing the other Island URLs, the agent providing the relay had
already shut down. This commit adds a keep_tunnel_open_time of > 5 to at
least get past that Island query timeout period.