Agent: Reimplement get_interface_to_target()
`get_interface_to_target()` has a few deficiencies:
1. It uses strings instead of more specific/useful types from the
   ipaddress module
2. On Windows, it sends UDP packets, which is noisy and hacky
3. It has some typing errors
4. It's too large (not addressed in this PR)

I believe that over 99% of the calls to this function involve a machine
and target on the same subnet. We can simply loop through the machine's
interfaces and check to see if the target IP is a member of any of the
machines networks without the need to send UDP packets or perform
complex socket operations.